# govctl_cli/api/openapi.yaml
openapi: "3.0.3"
info:
  title: "SoloCorp OS — Admin API"
  description: |
    REST API for SoloCorp OS governance, agent orchestration, pipeline
    monitoring, and system health. Used by the Dashboard UI and external
    integrations.

    **Base URL:** `http://localhost:8765`

    **Authentication:** API key via `X-API-Key` header (required for write ops).
  version: "1.0.0"
  contact:
    name: "Software Architect (พี่ทรงศักดิ์)"
    url: "https://github.com/DrSoloDev/Lab-solocorp-os2.4"
servers:
  - url: http://localhost:8765
    description: "Local development server"
  - url: https://api.solocorp.local
    description: "Production server (future)"
paths:
...
  /api/v1/agents/{id}/run:
    post:
      operationId: run_agent
      summary: Run an AO agent asynchronously
      description: |
        Invokes an AO agent with the provided context. The execution is
        async — the endpoint returns immediately with a trace_id, and the
        result can be polled via GET /agents/{id}/status?trace_id=...

        The request is published as an AO_REQUEST message to Central Bus,
        where the AO Bridge picks it up and spawns the agent.
      tags: [Agent]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            enum: [ceo, orchestrator, architect, engineering, qa]
          description: "Agent identifier"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRunRequest"
      responses:
        "202":
          description: "Agent execution accepted — async"
          headers:
            Location:
              schema:
                type: string
              description: "URL to poll for agent status"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/AgentNotFound"
  /api/v1/agents/{id}/status:
    get:
      operationId: get_agent_status
      summary: Poll agent execution status
      description: |
        Returns the current status of an agent execution identified by
        trace_id. Status transitions: pending → running → success|failed.

        Once status is 'success' or 'failed', the response includes the
        agent's output.
      tags: [Agent]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "Agent identifier"
        - name: trace_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: "Trace ID from the run response"
      responses:
        "200":
          description: "Agent execution status"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===================================================================
  # Pipeline
  # ===================================================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key for write operations. Configurable in gov/config.toml"
  schemas:
...
    AgentRunRequest:
      type: object
      required: [context]
      properties:
        context:
          type: object
          description: "Agent context — forwarded to build_prompt()"
          properties:
            project_id:
              type: string
            phase:
              type: string
            task:
              type: string
            input_data:
              type: object
            constraints:
              type: array
              items:
                type: string
        timeout:
          type: integer
          description: "Override default timeout in seconds"
          default: 300
    AgentRunResponse:
      type: object
      required: [trace_id, status, check_url]
      properties:
        trace_id:
          type: string
          format: uuid
          description: "Trace ID for polling and audit trail"
        status:
          type: string
          enum: [pending, queued]
          description: "Initial status (always pending on acceptance)"
        check_url:
          type: string
          format: uri
          description: "URL to poll for final status"
        agent_id:
          type: string
        message:
          type: string
          description: "Human-readable confirmation"
    AgentStatusResponse:
      type: object
      required: [trace_id, status]
      properties:
        trace_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, success, failed]
        agent_id:
          type: string
        elapsed_ms:
          type: integer
          description: "Elapsed time in ms (only when status is success/failed)"
        output:
          type: object
          description: "Agent output (only when status is success/failed)"
          properties:
            summary:
              type: string
            detail:
              type: object
            raw:
              type: string
        error:
          type: string
          description: "Error message (only when status is failed)"

    # ── Pipeline ─────────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: "Invalid request parameters"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: object
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: "RESOURCE_NOT_FOUND"
                  message:
                    type: string
...
    AgentNotFound:
      description: "Agent not found in registry"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: "AGENT_NOT_FOUND"
                  message:
                    type: string
                  details:
                    type: object
                    properties:
                      available:
                        type: array
                        items:
                          type: string

# ===================================================================
# Tags
# ===================================================================
...